+++ src/config/config.ts
@@ -2,6 +2,7 @@ import * as deepmerge from "deepmerge";
 import * as fs from "fs";
 import * as path from "path";
 import { Source } from "../context";
+import { NestedError } from "../error/nested-error";
 import { MetaTable } from "../meta";
 import { ElementTable } from "../meta/element";
 import { RuleConstructor } from "../rule";
@@ -196,7 +197,14 @@ export class Config {
 	public transform(filename: string): Source[] {
 		const transformer = this.findTransformer(filename);
 		if (transformer) {
-			return transformer.fn(filename);
+			try {
+				return transformer.fn(filename);
+			} catch (err) {
+				throw new NestedError(
+					`When transforming "${filename}": ${err.message}`,
+					err
+				);
+			}
 		} else {
 			const data = fs.readFileSync(filename, { encoding: "utf8" });
 			return [